adgm-corporate-agen/requirements.txt
streamlit
python-docx
langchain
langchain-openai
openai
unstructured[docx]
python-dotenv
faiss-cpu